smurf: pol2map: Use mosaic for pointing correction rather than first map
If a ref map is supplied by the user, it is used to define the pointing
corrections for each individual observation. But if no ref map is
supplied, the pointing corrections were previously determined by
comparing the map made from each observation with the map made from the
first observation. This is sub-optimal as the first observation may be
bad. So instead postpone the determination of the pointing correction
until after all observations have been mapped, and then use the mosaic of
all observations to determine the pointing corrections.