ary: Fix bug in annulling ary identifiers
The prologue of ary1_cpy.f says that the "loc" argument is "Given" - it
makes no mention of hte supplied value being changed. But in fact it can
be, if a primitive array is created rather than a simple array. On the
basis of the prologue the original C code used an "HDSLoc *" for the
argument (instead of an "HDSLoc **"), causing the new locator to be
discarded when the function exits, rather than being returned to the
caller. This left the caller with the original placeholder locator rather
rthan the new Array locator, causing all sorts of problems.

This is a bit worrying, because similar issues may exist in other places.
I relied on the F77 prologue docs to decide whether each function parameter
should be passed by value or by reference. I wish there was some warning
in gcc to indicate that a function parameter has been assigned a value...