hds: Include all required libraries in link script
The original link script for hds wrapper did not include hds-v4, hds-v5
and HDF5 libraries (explicitly or implicitly), but relies on the each
sharable library knowing its own dependencies. This was OK for linking
starlink apps, but caused problems for foreign language interfaces such
as perl and java. These use the hds_link script to determine the
libraries that must be included in the foreign language object.