smurf: Get the length of the JCMTSTATE arrays from an array, not a scalar
The JCMTSTATE extension in time-series models dumped by makemap begin
with the FTS_POS value, which is not always an array (e.g. if FTS not in
use). So always use RTS_NUM to define the length of the JCMTSTATE
arrays, rather than just using the first component in JCMTSTATE.

SMURF can create time-series data in which the