smurf: skyloop: Store models etc in temp directory, not current directory
The cleaned data, EXT, NOI and LUT models used to be stored in NDFs
within the current working directory. This made it difficult to see the
other files when looking at an "ls" listing, and cluttered up the place
in general. These NDFs are now stored in a subdirectory within the NDG
temporary directory.