pcs: Some more places where short file paths were assumed
The most important is the path to the adam directory in sup_deact. If
this was onger than 80 characters, an atask would abort with an error
when trying to update the GLOBAL parameters at the end of its execution.

You can get very long ADAM_USER values when using the Python scripts in
SMURF.