SC21: Fixed various typos, cross-refs, and performed general cleanup
In this commit, just before the starlink release in March 2018, I fixed a
variety of issues which I summarise here:

1. I added the acronym JSA = JCMT Science Archive to the acronyms page.

2. In section 3.1, how the DIMM works, I provided a mathematical description of
the signal received by a bolometer, following Chapin et al 2013.

3. I moved Figure 3.3 much closer to its first mention. It was separated by a
significant distance and I had trouble finding it/finding my way back while
reading through the documentation.

4. In section 3.7.4, the document said: "Here ast.zero_snr is used to constrain
the AST model to zero wherever the S/N is lower than 5?" but in the example
dimmconfig given, ast.zero_snr is set to 3 and ast.zero_snrlo is set to 2.
The example is the more correct way to run this dimmconfig file
(with ast.zero_snr = 3 and ast.zero_snrlo=2, that will give you better results),
so I updated the text.

5. The Data Chunking text box was originally defined as a figure with a caption
that simply said "Figure 3.2:". I changed this such that the label points to
the text box but the document does not refer to the text box as a figure. In
addition, this text box was referred to in Chapter 4 as the "text box below",
even though it was in the preceding chapter.

6. I also made more clear the assumptions that went into several statements about
pixel size, smoothing, and running ORACDR which were all specific to 850
microns and would not have worked for 450 micron data.

7. The appendix talking about SCUBA2_MAPSTATS was out of date and missing
descriptions of 3 columns generated by the function.

8. In addition, I did a thorough readthrough and fixed many typos throughout all
the chapters as well as in the appendices (including one in the name of the
jsa_generic dimmdonfig file), and added clarifying statements about group files
and their naming scheme.

9. I also fixed broken/incorrect links/references. Chapter 9 has the label
"sec:raw" while Section 2.3 has the label "sec:rawdata" and this was confused;
in Chapter 4, the SCUBA2_CHECK_RMS xreference was broken, etc.