smurf: pol2map: Allow differenes in observation FCFs to be found and used
The pol2map script now finds a calibration correction factor for each
observation by comparing the map made from the individual observation to
the coadd of all observations. This uses kappa:normalize with the offset
forced to zero, and restricted to the areas inside the AST mask. The
calibration correction factor is stored as FITS header CHUNKFAC in each
individual observation's map.

If the new parameter NORMALISE is set TRUE when running pol2map, any
calibration correction factors found in pre-existing maps are used to
scale the time-series data when creating maps from the data. This
approach (rather than just scaling the maps after they have been made)
means that the normalisation is effective whether using skyloop or
makemap to create the maps. If NORMALISE is left at its default value of
FALSE, any pre-existing calibration correction factors are ignored.

The idea is that the correction factors are found at the end of "step 1"
by comparing each auto-masked map with the coadd of all auto-masked maps.
They are then used to scale the raw data during the later steps that
create the external-masked maps (but only if NORMALISE is set TRUE when
initiating these later steps).

The object of this change was to reduce the dispersion between maps made
from different observations, and thus reduce the "MAPVAR=YES" variances
in the final catalogue. Sadly, using NORMALISE=YES seems to make very
little difference to the dispersions using my test set of 21 observations
of Serpens main field1.