KAPPA: Obtain the baseline-fit residuals about the median, not the mean.
Where the emission extends over a substantial fraction of the
spectrum, and is strong, such as in the Galactic Centre, the mean will
lie outside the baseline levels, and even heavy clipping can leave it
emission that is consequently not masked and biases the baseline fit.
In most cases the difference will be negligible.