ary: Prevent error when remapping arrays with different access modes
The original C version of ary1Imp reported an error if the requested
array was already mapped with a different access mode. This differed from
the F77 version (ary1_imp.f), which promoted any existing read-only DCB
entries to UPDATE access in such cases. This difference was introduced as
an attempt to reduce the chance of errors occurring in mult-threaded
applications. However, there are cases where the original behaviour is
required (e.g. kappa:wcsalign needs it when attmepting to modify the
input NDFs - when parameter INSITU is set TRUE). Further, experience from
smurf suggests that it should be the responsibility of the application
code to avoid simultaneous reading and writing of the same parts of an
array. And the original C behaviour was flawed anyway, because it did not
preclude the simulataneous mapping of an array for write access by two
threads, which is just as likely to produce threading issues as
simultaneous reading and writing.

So this change re-instates the original F77 behaviour.