hds: Provide F77 wrappers for HDS object locking functions
This is because F77 libraries (such as IRQ) are called from
multi-threaded contexts (such as smurf). This is thread safe because the
C wrappers for such F77 libraries use a mutex to ensure only one thread
is running an F77 function at any one time. However, the HDS object
locking scheme provided by HDS V5 still checks that the running thread
has a suitable lock on the object each time an F77 function uses an HDSV5
locator. So F77 requires a way to lock and unlock HDS locators.