smurf: pol2map: Ignore low weight observations
New parameter WEIGHTLIM specifies the lowest weight for usable
observations, in the range 0.0 to 1.0 (default 0.05). Observations that
have weights below this value are excluded from the coadded I, Q and U
maps and the vector catalogue. The weights are based on the similarity of
the observation's auto-masked map and the coadd of all auto-masked maps.

Since the excluded observations have low weight, they make little
difference to the final producrs if they are included. The real reason
for excluding them is because they can cause makemap to crash.