smurf: calcqu: Fix bug that caused sources to be smeared out
Resample the JCMTState info using the correct scaling factor.
Previously, the scaling factor did not take account of the fact
that some input time slices are not used. This resulted in the
wrong pointing info besing assigned to samples, with an
increasing error along the time stream, resulting in the source
being smeared. The degree of smearing was related to how many input
slices were unused and so varied from observation to observation.