smurf: calcqu: Fix bug that caused blurring of some observations
Yesterday's fix cleared up cases of really bad scaling - e.g. where a
potentially large section of the time stream was ignored because of jumps
in the POL_ANG value. But it leaves smaller scaling issues in a
significant number of observations. This can cause makemap to create
badly blurred images The problem was that the JCMTState resampling code
was assuming that all boxes are the same length, which is not the case.
So now, instead of using smfDownsamp1, each important JCMTState value in
the output is calculated within this function as the average of the input
JCMTState values for the time slices that contibute to each output I, Q
or U value.