KAPPA: Update version numbers and delineate releases.
The normal procedure is to have a decimal release at each Starlink
release, but this time as there was no such increment.  So instead of
2017A being Version 2.5 and 2018A being Version 2.6, the latter has
Version 2.5-9.  The -8 -> -9 increment was for the new parameter
in SETAXIS that was forgotten in Commit 15ad51ccf5c9.

This separation between 2017A and 2018A, forces a split of the
respective developments in the help and release notes.