ndf: New tuning parameter to fix arbitrary date/time in new history
The new tuning parameter NDF_FIXDT is intended to facilitate regression
testing, where the difference in date/time in history records could cause
a test to fail. If 1, the date and time string stored in any new History
records will be fixed at an arbitrary value ("10-AUG-2018 11:00:00")
regardless of the current time. If 0 (the default), the real date and
time will be used.