gaia: (polarimetry toolbox) allow all vectors to be drawn with unit length
If the vector scale value is negative, all vectors are now drawn with a
fixed length equal to the absolute scale value (in image pixels). This is
to satisfy a user request for a way to draw all vectors with the same
length. It would have been better to add a "constant" item to the list of
options under "Vector length", but that would have taken way too long to
implement, given it is only one person who has asked for this feature.
Doing it this way was much much faster. It would not have happened using
the (better) alternative approach - cest la vie.