kappa: hiscom:  Allow application name to be specified
New parameter APPNAME can be used to change the application name stored
in the new history record from the default of "HISCOM". If APPNAME is
set, the default history that records the parameter values supplied to
HISCOM is not written. So the new history record then contains just the
date, application name and formatted text supplied to HISCOM.

Scripts that generate NDFs can use this facility to record the details of
the invocation of the script in the form of a history record in the
output NDF. ORAC-DR does the saem thing by calling the Perl interface for
NDF_HPUT directly.