smurf: pol2map etc: Do not display ndfecho info but default
The NDG class in starutil uses kappa:ndfecho a lot to check for the
existence of NDFs. This can clog up pol2map log files with harmless error
messages and lists of input files. This change causes such information to
be displayed only if the script is run with parameter MSG_FILTER=DEBUG.