smurf: makemap/skyloop - fix issues caused by 5c1ad43216
5c1ad43216 modified smf_concat_smfGroup so that the FITS header
associated with the concatenated time stream did not include any headers
that have different values in the individual time streams (previously it
arbitrarily adopted the values from the first individual time stream).

However, it seems that makemap relies on some value being available in
the concatenated time stream for certain headers that are now deleted due
to them not being constant across all individual time streams (NSUBSCAN
and STEPTIME).

This commit adds checks to ensure that these required keywords are
present and if not, handles the errors appropriately.