smurf: makemap: Modification to the ast.filt_diff facility
ast.filt_diff removes the low frequency changes introduced into the
map at the end of each iteration. The spatial size of the filter is
determined by the value of ast.filt_diff, which in practice would
normally be set to half the value of flt.filt_largescale. This change
causes the low frequencies to be scaled to fit the original map before
removing them from the map. It doesn't seem to make a big difference to
the effects of using ast.filt_diff.

ast.filt_diff is supposed to be an alternative to AST masking. However,
not suprisingly, the background regions are not so flat using
ast.filt_diff as using AST masking. An option would be to use both - AST
masking to flatten the background regions and ast.filt_diff to reduce
spurious large-scale features inside the AST mask.