esp: ellpro: Use assumed size array declarations
Various subroutine dummy arguments are declared with fixed sizes that are
inappropriate to the way in which the subroutines are called. This was
not a problem with older versions of gfortran, but gfortran 8.2.1 (as
distributed with, for instance, Fedora Core 28) objects.  So this commit
changes these dummy arguments to use assumed-size declarations.