kappa: ardplot: Avoid distortion when drawing large regions
Change the tangent point for the projection used to display regions
if there is no underlying picture. The tangent point used to be the
south east corner of the Region's bounding box. Now it is the centre
of the bounding box. This is because Regions that cover a large
fraction of the sky were being displayed with lots of distortion.