hds: Ensure deferred error reporting is used
The selection of v4 or v5 code is in some cases done by calling the v4
code and checking for an error, in which case the error is annulled and
the v5 code is called. This relies on deferred error reporting, so that
the error can be annulled before it is reported. Calling code will
usually have established a deferred error reporting context before
calling HDS, so all is usually ok. But some code (it seems like maybe the
NDF perl module for instance) does not establish a deferred error
reporting context, and so the error reported by the v4 code is delivered
immediately, causing confusion and worry.

This change ensures that all such V4 code is executed in a deferred error
reporting context, regardless of whether such a context has been
established by the calling code, or not.