kaplibs: Read AST objects directly from FITS files
The kpg1_gtobj routine is used by applications to get an AST object
from the user. The user can supply the object in several forms:

1) as an HDS character array holding a textual dump of an arbitrary AST object

2) as an NDF (the WCS FrameSet is returned to the application)

3) as a text file containing a textual dump of an arbitrary AST object,
or a set of FITS WCS header.

This change adds a fourth option - to get an AST object from a FITS file.
If the file spec supplied by the user contains an extension (e.g.
"myobj.fit[2]", etc), then only that one extension is checked. Otherwise
all extensions are checked (starting with the primary HDU), until an
extension is found from which a usable AST object can be obtained. At the
moment, this is limited to two cases:

a) reading a WCS FrameSet from the header of an image extension (or
primary HDU)

b) reading a MOC from a binary table extension

So for instance, with this change it is possible to draw the outline of a
MOC held in a FITS file over the top of another image using kappa:ardplot:

% display fred mode=current
% ardplot region=myarea.fit

Also, ATOOLS commands that expect an AST Region can nw be supplied with a
FITS file. For instance,

% atools
% astoverlap myarea.fit circle.ast

(where for instance the file "circle.ast" contains a dump of an AST
Circle object).