smurf: starutils: Fix bug removing path from full file spec
Removing the current working directory path from "/a/bc/d.sdf" would fail
if the current working path was "/a/b", leaving "c/d.sdf" as the relative
path. The correct action, since "/a/bc/d.sdf" does not start with the
current working path "/a/b", is to leave the absolute path unchanged.