smurf: Add info to tausrc=auto error message
In order to help distinguish different types of WVM problems,
the final percentage of good samples has been added to the message,
along with the percentage prior to despiking.  Additionally a
descriptive sentence has been added to indicate whether the
WVM data are missing (< 1%), insufficient (< 80%) or
insufficient after despiking.

In order to add the extra information, some fixed text had to be
removed in order for the message to fit within the size constraint.