smurf: Add WVMLOG param. to control WVM debug o/p
This is to allow the WVM debugging output to be seen more easily
than recompiling with #define DEBUG_WVM.  When the despiking and
smoothing steps are active, the WVM data are first copied to a
temporary array.  The three sets of data (original, despiked and
smoothed) are then written in a single table and the temporary
arrays freed.  The column names refer to the process just completed,
so, for example, if only despiking is done, the "TAU" column contains
the original data (pre-despike) and the "TAU_DESPIKE" column contains
the final data (in the main taudata array).