smurf: Correct sense of rts_end comparison
It looks like this loop is supposed to skip over values that aren't
increasing, so it should run while they are mis-ordered.  The loop
is still somewhat odd in that it starts at islice=1 and looks back at
islice-1 and then does the calculation looking forward at islice+10.
So it will never use the first and 11th values (islice=0) even if the
data are good from the start.