smurf: makemap - avoid strange features in variance if using pca
The variance array for daisy maps created with the PCA model can show
rings of extremely low variance. This seems to be due to the PCA model
removing almost the entire signal from some bolometers. This happens for
bolometers that are "unique" - i.e. have almost nothing in common with
other bolometers. For each such bolometer, one of the principal
components will end up being nearly identical to the bolometer signal
(since each bolometer must be a linear combination of one or more PCA
components) and so leave no significant residuals after removal of the PCA
model.

This change gets round this by flagging bolometers as bad if the removal
of the PCA model causes the noise to drop too much.