smurf: Improve PCA convergence
If a bolometer is flagged as entirely bad by the PCA model because its
noise drops by too much, the flag is now permanent. Previously, the flag
was cleared on the next itreration and the bolometer tested again to see
if the noise stills drops too much. But this can cause border-line
bolometers to toggle too and fro between being rejected and not rejected.
This prevented convergence.