smurf: pol2ip: Override default for new pca.noiselim config parameter
The pca.noiselim parameter default to 0.1, which is fine if the common
mode has been removed before PCA is used (as in pol2map). But if the
common mode is still present in the data (as in pol2ip), it's removal by
the PCA model looks like a huge reduction in noise, which causes all
bolometers to be rejected if the default value for pca.noiselim is
retained. SO instead set pca.noiselim to zero so that no bolometers are
rejected because of too much noise removal.