perlmods: Update NDF module
This updates the NDF module to the tip of the "ndf_c" branch,
where it uses the C interface to NDF (and HDS, etc.).

ExtUtils::F77 has been removed from cpan.deps as it is no longer
required to build the NDF module.  Astro::SLA has been removed as
it requires ExtUtils::F77 but is no longer required itself.