latexsupport: Switch back to the original naming for linked packages.
After changing my mind a few times, I decided to stick with the
original naming scheme of \PACKAGEref to obtain a link to a package
with the package name in small capitals.  The \package style is
relatively recent, mostly in the cookbooks related to JCMT
instruments.  Also there needed to be exceptions because of clashes
with LaTeX markup (\ast and \kappa).

It is therefore less work to standardize the documents that broke with
the informal convention than the reverse.

Also removed extra brace from moved \oracdr and \topcat deprecated
definitions.

Noted the duplicated deprecated \oracdr definitions and chose the
the one that works and commented out the other.