SC/4: Improve file operators layout of minipage.
This improves the PDF formatting, but not for the HTML where the text
and first example appear to the right of the table instead of below.
There are a few minipages in SC/4 that show formatting problems.
Unleess tex4ht improves its handling of minipages, we have to decide
whether the PDF or HTML format takes priority, since we're not allowed
to have latexonly and htmlonly blocks.