smurf: pol2check: Coadded maps have no UTDATE or OBSNUM headers
The key returned by pol2check for any supplied, I, Q or U maps was
assuming that the maps were made from a single observation and so still
had UTDATE and OBSNUM headers. But a coadded map created from multiple
observations will not have these headers. So instead use a a key of the
form "<OBJECT>_<DATE-OBS>" for coadded maps. The key will be the same for I
Q and U coadds and so can be used to associate them together, as required
by pol2map.