smurf: pol2map - Use default CHUNKFAC value if pointing correction fails
If step 1 fails to find a pointing correction for an observation (e.g.
the observation's auto-masked I map cannot be aligned with the reference
map), then ensure a default value of 1.0 is stored for the observations
CHUNKFAC header. Previously, no value would be stored for CHUNKFAC,
causing step 2 to fail with a python error. Now, if step 2 fails to find
a CHUNKFAC value, it will issue a warning and continue using a default
value of 1.0.

cannot for an auto-masked I map created at step 1 for an observation cannot be aligned with the

If pointing correction fails at step 1 for an observatio