smurf: makemap - ensure no output map is created if an error occurs
Previously, a valid output NDF was still created if certain errors (such
as "Too few samples remain") occurred. In such cases the output NDF is
useless (filled with zeros) and its creation can mask the fact that an
error has occurred, potentially confusing scripts. So this commit ensures
that no output NDF is created if an error of any type occurs.