smurf: polpack - obs maps are not required for IPREF image
If an output catalogue is being created, but IOUT is null, then
the supplied IPREF map is used as the total intensity map when
creating the vectors. In this case, do not check that the individual
observation I maps exist since they have not been requested and
may not be available. Previously such checks caused pol2map to
abort if any of the individual observation I maps were missing.