atl: Preserve end of line white space in MOC source function
End of line white space is significant in the MOC string serialisation,
so preserve it. In fact it is not possible to preserve it since GRP
strips the carriage return/line-feed at the end of each line. Instead,
ensure every line returned by the source function ends with a space,
which will have the same effect as the missing carriage return/line-feed.