atl/kaplibs: Allow AST objects to be read from files with very long lines
Previously, KPG1_GTOBJ used the GRP library to read and store text from
files. However, GRP truncates lines of text at 255 characters. So
KPG1_GTOBJ could not read objects from files with long lines. This commit
chanages KPG1_GTOBJ soi that it uses a new facility within ATL to read
and store lines of text instead of GRP. This new facility is a simple
scheme that uses a C structure to hold multiple lines of text with
arbitrary lengths. It is referred to in the code as a "virtual file
structure" (VFS). The main routines for accessing and creating VFSs are
in atl2.c. If this proves to be of wider use, it should be moved out of
ATL into its own library.