kappa: Add initial version of COMPLEX program
This program is intended to operate on complex NDFs, e.g. to extract
the real and imaginary parts, compute the modulus and argument or to
convert combine these parts into a new complex NDF.  The operation
of the program is determined by its INTYPE and OUTTYPE parameters.

This program is based on MAKESNR and similarly does not make use
of the LPG library.  It also does not yet deal with the variance
array or take noise into account in its calculations.  It assumes
that the given argument NDF is in radians.