smurf: pol2map - fix 3-step processing
This is a fix for a bad bug that caused the output catalogue values to be
completely wrong if a previously created coadd was reused AND a binned-up
output catalogue is being created (i.e. CAT is non-null, BINSIZE is
non-null and one or more of IOUT, QOUT and UOUT is null). The problem was
that the existing coadd was not binned up prior to creating the
catalogue, so the association between corresponding (I,Q,U) values was
lost. In other words, the wrong set of I, Q and U values were being
combined together to create each pair of (P,ANG) values.