cupid: More multi-threading bug fixes
Note the multi-threaded calculation of the defrault RMS value is never
going to be *exactly* the same result as the old code. This means that
you need to specify the RMS explicitly rather than taking the default if
you want to compare old and new results. You also need to specify
"flatslope=0" to compare results since whether a peak is on the coastal
plane or not may occasionnaly depend on which way you approach it (i.e
whether the route is very long with low gradient - below flatsope - or
very short with high gradient).