smurf: pol2map: avoid re-creating all I maps unnecessarily
If an existing I coadd is supplied (either explicitly or as the IP
reference map), and an output catalogue is being created, and BINSIZE >
PIXSIZE, and MAPVAR=yes, then the individual obs I maps are needed (they
are binned up to BINSIZE before being coadded to form the I coadd used to
create the catalogue). If they already exist then they are re-used (if
REUSE=yes). Otherwise, they need to be re-created from the I time-series.
The test to see if the individual I obs maps already exist used to test
for I obs maps that cannot be created. For instance, if an observation
was split into two chunks - one long and one very short - then it will
probably not be possible to create an I map from the short chunk, so the
test should not look for such an I map. The previous test included such
short observation chunks and consequently failed in such a situation,
causing all the observation I maps to be re-created unnecessarily. In
fact there is no way that I can see to get a list of the expected I obs
maps that is 100% reliable. So the new test just checks that I obs maps
exist for more than 50% of the observation chunks. In practice, it's
probably going to be a case of all exist or none exist, so this should
usually be safe.