smurf: New command 'matchbeam' to smooth 450 maps to 850 resolution
It determines a suitable smoothing kernel based on the two-component beam
shapes defined in Dempsey et al 2018. It then uses this kernel to smooth
a supplied 450 um map. Optionally, the 450 map may first be resampled to
match the pixel grid of a supplied reference map.  This is useful if you
want to modify a 2" 450 map so that it can be compared pixel-for-pixel
with a 4" 850 map.