smurf: makemap: Normalise IP ref map to observation before using it
If it is known that the IP reference map has a different calibration to
the observation being processed, then the IP reference map should be
scaled to the calibration of the observation before being used.

The relative calibration factor is supplied in the CHUNKFAC config
parameter. It is set by pol2map in step 2/3, if normalise=yes, based on
a comparison of the I map made for the observation at step 1, with the
coadd of all I maps.

This change will affect the IP correction and thus the final Q and U
maps if a value is set for the chunkfac config parameter (i.e. only if
normalise=yes when running pol2map at step 3).