smurf: pol2map: set normalisation factor to 1.0 for badly correlated maps
Previously, maps for single obervations that were not well correlated
with the coadd of all observations could result in somewhat spurious
normalisation factors, which caused the final mapvar=yes variances to
increase rather than decrease (if normalise=yes). This changes causes the
normalisation factor to be set to unity for any observation that has a
correlation coefficient (when compared to the coadd) smaller than 0.95.
This figure is a bit arbitrary but seems to work reasonably well for NGC
6334 field 2. Hopefully Ray will be able to check it out on a wider range
of fields.