KAPPA: Increment version number.
Although Commit acf808e4b649f does not introduce a new parameter, it
does introduce a new value.  In ORAC-DR it is sometimes necessary to
inquire the version number of applications, especially KAPPA, to ensure
that a new facility is not accessed while running an older Starlink.
To test whether or not xmap=LRLinear can be used, a new version is
needed.