kappa: speed up COLLAPSE by multi-threading popular estimators
This commit multi-threads the MEAN, WMEAN and SIGMA estimators. It also
increases the max block size from 8 mega-pixels (set in 2006 when most
desktops had 0.5-1 GB of memory) to 80 mega-pixels.

This combination gives a typical speed-up of between a factor of 4 and 5
when collapsing big arrays. This is assuming the input array is already
in the memory cache, as it will usually be if another kappa app has just
processed it. If this is not the case, it can take far longer to read the
array into memory than it does to collapse the array (at least on my
machine).