spt: Add starf2c - a basic tool for helping with F77 to C conversion
Reads an F77 source file in standard StarLSE format and writes vaguely
equivalent C code to standard output.

There are probably many bugs, unsupported Fotran patterns, etc, in this
script. The main reason I've added it to SPT (Software Porting Tools) is
so that it does not get lost. How much practical use it will be remains to
be seen. It was originally used to convert the NDF library from F77 to C,
and still contains much baggage that is not relevant outside of that
task. It is intended to do nothing more than take some of the donkey work
out of doing an F77 to C conversion by hand. In nearly all cases the
resulting C code will need to be examined carefully and changed in many
ways. In particularly, character handling in Fortran is very different to
C, and the resulting C code will nearly always need changing if the code
involves anything other than simple assignments. Also, array indices in
the C code should be checked carefully.

A temporary version of the Fortran source file should be created before
using this tool which adheres to the following requirements:

- All uses of "A**B" should be replaced by "POW(A,B)"

- Generic source file should use "<TYPE>" to represent the data type in
the prologue and not "?".