kaplibs: New function kpg1SymFit to do symetric linear fit
By "symetric", it means that fiiting X to Y produces the algebraic
inverse of fitting Y to X. This is not true for the usual form of least
squares fitting, which minimises Y deviations assuming the X values are
exact. This new function is appropriate for cases where there are random
errors in both X and Y values.