polpack: Allow debiasing to be done with modified asymptotic estimator
polbn and polvec now have an extra parameter called DEBIASTYPE, which can
be used to select the debiasing method. In addition to the old asymptotic
estimator, there is now an option to use the "modified asymptotic
estimator". See https://arxiv.org/pdf/1407.0178.pdf section 2.5.

The benefit of this estimator is that it can be evaluated for all (Q,U)
values. This is unlike the traditional asymptotic estimator (see section
2.3 of the aboe paper), which can only be evaluated for (Q,U) values for
which Q**2+U**2 >= variance.