kappa: scatter: improve linear fit
The linear fit is based on the points that fall within the bounds of the
plot. But the bounds of plot may exclude more points on one side of the
line than the other, introducing a bias to the fit. This change attempts
to remove this bias by also ignoring points if they are outside the
bounds of the plot after being reflected through the current best fit
line.