prm: Add extra vec?2 functions to solve OSX build problem
The ARY library has cgeneric functions that, before optimisation, refer
to vec "integer 2 <T>" functions such as vecI2D. Since the input to such
functions is already integer, no rounding is needed and so the "2"
functions are never called and the compiler (on Linux) optimises out the
calls. But on OSX the compiler does not do this, requiring the unused
functions to exist in order to avoid a linker problem.

This changes adds the unused "vec?2<T>" functions. These just call the
corresponding "vec?to<T>" functions.