kappa: new command PIXBIN bins pixels in a generalised manner
It places each input pixel into an output bin and returns a new NDF
holding all the input values that fall in each bin. The bin for each
input value is specified by one or more index NDFs. For those who are old
enough, it's a bit like edrs:pixmap except that it stores all the input
pixels in each bin individually in the output instead of storing their
mean. The idea is that you will probably then want to use COLLAPSE to
collapse the contents of each bin into some value representative of the
bin (e.g. the mean or what ever).