smurf: pol2noise - allow noise to be smoothed without loosing resolution
The new "Re-model" mode allows the noise estimated in a POL2 vector
catalogue to be re-modelled, effectively smoothing them without loosing
resolution.

Also, the "Display" mode has been changed to incorporate two different
and independent methods for estimating the relationship between the noise
estimates stored in the error columns and the noise in the data columns.
Sadly, neither seems to be particularly reliable.