starconf: gcc 10 breaks Starlink use of %VAL
As of gcc 10, gfortran reports an error by default rather than a warning
if multiple calls to the same subroutine have different argument types.
This break most uses of %VAL in starlink code, which is always assumed by
the coimpiler to have a type of INTEGER*8.

This commit fixes this problem by adding the -fallow-argument-mismatch
option to STAR_FCFLAGS and STAR_FFLAGS, if the option is supported by the
Fortran compiler.