extreme: avoid multiple definitions of global variables
As of gcc 10, -fno-common is the default instead of -fcommon. This causes
errors to be reported because extereme does not use the extern keyword
in a standard manner when declaring global variables.

This fix includes "extern" keywords as required to turn most definitions
into declarations. The definitions are now produced when ygen.h is
included from ygen.c.