hds: hdsNew wrapper now ensures the new file is not already open
If the file to be created by hdsNew is already open within hds - either
V4 or V5 - an error is now reported. The versions of hdsNew provided by
the V4 and V5 libraries each check to see if the specified file is
currently open within the same library, but cannot check if the file is
open in the other library. So the wrapper for hdsNew does this extra
check, using the implementation of the hdsIsOpen function provided by
each library.