ndf: F77 API - use a single space C string for a blank F77 string
Previously, if an F77 string full of spaces was supplied to the F77
wrapper for an NDF function, it would be converted by cnfCreim into a
null C string. But a null C string has length zero, which can cause
problems (e.g. trying to create a _CHAR*0 will fail). Now, such an F77
string is converted into a C string holding a single space, and so will
be represented by a _CHAR*1 holding a single space.