KAPPA: Make centre unit clearer for sky-offset co-ordinates.
There is already a heading giving the type and units of the centre
co-ordinates.  This was to avoid long or superfluous information, such
as obvious sexagesimal positions, detracting from the numerical
answers.  However, for offset positions, there is potential for
ambiguity, namely that the unit is arcsecond for both co-ordinates and
their errors, whereas the positions are in degrees.

As a compromise, the unit is reported after the co-ordinates if
it is the same for both WCS axes.